Add a reminder to clear cache when static figures are updated
When a static figure is changed, the showyourwork build on github does not take the update into account right away because it uses a cached version. This can lead to the situation where an old figure ends up in the PDF even though the updated version is in the `src/static/` directory.

I suggest to add a line in the docs to spare inexperienced users like myself the troubleshooting.